+++ visualization/app/codeCharta/ui/ribbonBar/ribbonBar.module.ts
@@ -5,6 +5,7 @@ import "../colorSettingsPanel/colorSettingsPanel.module"
 import { downgradeComponent } from "@angular/upgrade/static"
 import { SearchPanelComponent } from "../searchPanel/searchPanel.component"
 import { CustomConfigsComponent } from "../customConfigs/customConfigs.component"
+import { HeightSettingsPanelComponent } from "./heightSettingsPanel/heightSettingsPanel.component"
 
 angular.module("app.codeCharta.ui.ribbonBar", ["app.codeCharta.state"])
 
@@ -13,3 +14,4 @@ angular
 	.component(ribbonBarComponent.selector, ribbonBarComponent)
 	.directive("ccSearchPanel", downgradeComponent({ component: SearchPanelComponent }))
 	.directive("ccCustomConfigs", downgradeComponent({ component: CustomConfigsComponent }))
+	.directive("ccHeightSettingsPanel", downgradeComponent({ component: HeightSettingsPanelComponent }))